push

+++ Module_05/Event.hpp
@@ -6,8 +6,8 @@ class Event {
     private:
         std::string _reason;
         Node        *_location;
-        int         _probability;
-        int         _time;
+        float       _probability;
+        float       _time;
     public:
-        Event(std::string reason, Node *location, int prob, int time);
+        Event(std::string reason, Node *location, float prob, float time);
 };
