+++ Module_05/Rail.cpp
@@ -1,5 +1,5 @@
 #include "Rail.hpp"
 
-Rail::Rail(Node *a, Node *b, int lenght) : _a(a), _b(b), _lenght(lenght) {
+Rail::Rail(Node *a, Node *b, float lenght) : _a(a), _b(b), _lenght(lenght) {
 
 }
