+++ Module_05/Train.cpp
@@ -1,5 +1,5 @@
 #include "Train.hpp"
 
-Train::Train(std::string name, Node *a, Node *b, int acc, int brake, int departure) : _name(name) , _departure(a), _arrival(b), _maxAcceleration(acc), _maxBrakeForce(brake), _departureHour(departure) {
+Train::Train(std::string name, Node *a, Node *b, float acc, float brake, float departure) : _name(name) , _departure(a), _arrival(b), _maxAcceleration(acc), _maxBrakeForce(brake), _departureHour(departure) {
 
 }
