+++ Module_05/Train.hpp
@@ -7,9 +7,9 @@ class Train {
         std::string _name;
         Node        *_departure;
         Node        *_arrival;
-        int         _maxAcceleration;
-        int         _maxBrakeForce;
-        int         _departureHour;
+        float       _maxAcceleration;
+        float       _maxBrakeForce;
+        float       _departureHour;
     public:
-        Train(std::string name, Node *a, Node *b, int acc, int brake, int departure);
+        Train(std::string name, Node *a, Node *b, float acc, float brake, float departure);
 };
